+++ src/stations/StationsLoader.ts
@@ -1,8 +1,8 @@
 import zlib from 'zlib'
-import request from 'request'
+import got from 'got'
 import JSONStream from 'JSONStream'
 import { stationTransformer } from './stationTransformer'
-import defaultConfig from '../config/config.json'
+import defaultConfig from '../config/config.json' assert { type: 'json' }
 import { Station } from './schema'
 import { Readable } from 'stronger-typed-streams'
 
@@ -13,7 +13,7 @@ export class StationsLoader {
     const unzipStream = zlib.createGunzip()
     const { url } = this
     const { headers } = defaultConfig
-    return request({ url, headers })
+    return got.stream(url, { headers, decompress: false })
       .pipe(unzipStream)
       .pipe(JSONStream.parse('*'))
       .pipe(stationTransformer)